Sync upload form with Google Music DOM
The form was also changed to mimic the Trash listing rather than
the playlists, as it is similarly administrative in nature.
This unfortunately meant that there wasn't a good place to put
an input file picker button, so drag 'n drop is for now the only
interface for uploading songs.