Redesign options page, take device ID as a parameter
Requiring user input on a matter as esoteric as this is unfortunate,
but trying to capture the user's e-mail address from the Google Music
DOM is fragile and annoying to code.  To get this done clearly, the
options page had to be complicated considerably, but I think this is
still preferable.